[Attributor][FIX] Ensure stable iteration order
With
https://github.com/llvm/llvm-project/commit/f82b3ed81c8ee3815b35c24ff3d24770d46bdd6f
we introduced an ordering issue revealed by the reverse iteration
buildbot. Depending on the order of the map that tracks the AAIsDead AAs
we ended up with slightly different attributes. This is not totally
unexpected and can happen. We should however be deterministic in our
orderings to avoid such issues.