[InstCombine] Remove unnecessary byval check in callee cast fold
The logic for handling this was fixed in
e864cc13f253f12cc5edb917edd6b6629b410838, but the check for byval
on the callee was retained. This resulted in a weird situation
where the transform would work depending on whether the byval
was only on the call or on both the call and the function.