[ConstraintElimination] Remove over-eager assertion.
After moving the CanAdd check in 0a02d12dbcefc5a4cb2 and using it for
the assume cases as well, the passed in block may not have  a branch
instruction as terminator. This can trigger the assertion. Given the new
use case, it doesn't add value any longer and can be removed.

Fixes https://github.com/llvm/llvm-project/issues/54281